assignPhotoToStudent-mapping omgeschreven naar profilecontroller en -service (methodenaam nu uploadProfilePicture())
