+++ src/main/java/nl/novi/eindopdracht/controllers/ProfileController.java
@@ -1,10 +1,12 @@
 package nl.novi.eindopdracht.controllers;
 
+import nl.novi.eindopdracht.models.FileUploadResponse;
 import nl.novi.eindopdracht.models.Profile;
 import nl.novi.eindopdracht.services.ProfileService;
 import org.springframework.http.HttpStatus;
 import org.springframework.http.ResponseEntity;
 import org.springframework.web.bind.annotation.*;
+import org.springframework.web.multipart.MultipartFile;
 
 import java.util.List;
 
@@ -12,9 +14,11 @@ import java.util.List;
 @RequestMapping("/profiles")
 public class ProfileController {
     private final ProfileService profileService;
+    private final PhotoController photoController;
 
-    public ProfileController(ProfileService profileService) {
+    public ProfileController(ProfileService profileService, PhotoController photoController) {
         this.profileService = profileService;
+        this.photoController = photoController;
     }
 
     //MAPPINGS
@@ -32,6 +36,13 @@ public class ProfileController {
     public ResponseEntity<Object> updateProfile(@PathVariable String username, @RequestBody Profile newProfile) {
         profileService.updateProfile(username, newProfile);
         return ResponseEntity.ok().body("Je wijzigingen zijn opgeslagen");
+    }
+
+    @PostMapping("/{username}/photo")
+    public void assignPhotoToStudent(@PathVariable String username,
+                                     @RequestBody MultipartFile file) {
+        FileUploadResponse photo = photoController.uploadFile(file);
+        profileService.assignPhotoToProfile(photo.getFileName(), username);
     }
 
 }
