+++ src/main/java/nl/novi/eindopdracht/models/User.java
@@ -32,9 +32,6 @@ public class User {
             cascade = CascadeType.ALL)
     Set<Request> outgoingRequests;
 
-    @OneToOne (cascade = CascadeType.ALL)
-    FileUploadResponse fileUploadResponse;
-
     private String password;
     private String email;
     private boolean enabled;
@@ -59,14 +56,6 @@ public class User {
 
     public void setUsername(String userName) {
         this.username = userName;
-    }
-
-    public FileUploadResponse getFileUploadResponse() {
-        return fileUploadResponse;
-    }
-
-    public void setFileUploadResponse(FileUploadResponse fileUploadResponse) {
-        this.fileUploadResponse = fileUploadResponse;
     }
 
     public Profile getProfile() {
